requirements.txt
alerce==1.2.0 
astropy==5.3.3 
fastapi==0.103.2 
matplotlib==3.8.0
numpy==1.26.0 
pandas==2.1.1
psycopg2-binary==2.9.7
requests==2.31.0
SQLAlchemy==2.0.22  
uvicorn==0.23.2 